requirements.txt
langchain-openai
langchain-core
python-dotenv
feedparser
newspaper3k
nltk
